Add Correlation table and moving isx funtions to another file (#22)
* runs DI automatically new paradigms of step functions

* Update pipeline_functions.py

* Update pipeline_functions.py

* Update filehandler.py

* adds compute_traces_corr

* Update analysis_utils.py

* de interleave to functions,  get focus to aux

* projection as a function

---------
