+++ cgk_calcium_tools/filehandler.py
@@ -6,9 +6,9 @@ import warnings
 import numpy as np
 import isx
 import json
-from typing import Union, Tuple, Iterable
-from .processing import fix_frames
+from typing import Union, Tuple
 import shutil
+import pandas as pd
 from .files_io import (
     write_log_file,
     remove_file_and_json,
@@ -23,7 +23,10 @@ from .isx_aux_functions import (
     cellset_is_empty,
     create_empty_cellset,
     create_empty_events,
+    get_efocus,
 )
+from .pipeline_functions import f_register, f_message, de_interleave
+from .analysis_utils import compute_traces_corr
 
 
 def ifstr2list(x) -> list:
@@ -92,7 +95,7 @@ class isx_files_handler:
         outputsfolders: Union[str, list] = ".",
         data_subfolders: Union[str, list] = ".",
         files_patterns: Union[str, list] = "**/*.isxd",
-        processing_steps: list = ["DI", "PP", "TR", "BP", "MC", "DFF", "PM"],
+        processing_steps: list = ["PP", "TR", "BP", "MC", "DFF", "PM"],
         single_file_match: bool = False,
         recording_labels: Union[list, None] = None,
         check_new_inputs: bool = True,
@@ -104,7 +107,6 @@ class isx_files_handler:
     ):
 
         self.processing_steps = processing_steps
-        self.deinterleave_output_files = []
 
         # Check if the parameters file exists at the given path, open and load it.
         assert os.path.exists(parameters_path), "parameters file does not exist"
@@ -216,6 +218,7 @@ class isx_files_handler:
                             + "_metadata.json",
                         )
                         if os.path.exists(json_file):
+                            pb.update_progress_bar(1)
                             continue
 
                     # Read the video file and grab metadata information
@@ -381,27 +384,18 @@ class isx_files_handler:
         assert (
             operation in self.processing_steps
         ), f"operation must be in the list {self.processing_steps}, got: {operation}"
-
-        # Copy the list of processing steps to a local variable
-        local_processing_steps = self.processing_steps
-
-        # Remove DI from list, if its first
-        if local_processing_steps[0] == "DI":
-            local_processing_steps.remove("DI")
-
-        opi = local_processing_steps.index(operation)
+        opi = self.processing_steps.index(operation)
 
         outputs = []
         inputs = []
 
-        suffix_out = "-" + "-".join(local_processing_steps[: opi + 1]) + ".isxd"
+        suffix_out = "-" + "-".join(self.processing_steps[: opi + 1]) + ".isxd"
 
         if opi == 0:
             suffix_in = None
-            # inputs = self.p_rec_paths
-            inputs = self.deinterleave_output_files
+            inputs = self.p_rec_paths
         else:
-            suffix_in = "-" + "-".join(local_processing_steps[:opi]) + ".isxd"
+            suffix_in = "-" + "-".join(self.processing_steps[:opi]) + ".isxd"
 
         for ofolder, file in zip(self.p_outputsfolders, self.p_rec_paths):
             outputs.append(str(Path(ofolder, Path(file).stem + suffix_out)))
@@ -446,28 +440,30 @@ class isx_files_handler:
             outputs.append(str(Path(ofolder, Path(file).stem + suffix_out)))
         return outputs
 
-    def check_str_in_filenames(self, strlist: list, only_warning: bool = True) -> None:
+    def compute_traces_corr(fh, cellsetname: str, verbose=False) -> pd.DataFrame:
         """
-        This method verify the order of the files. Checking that each element of the
-        input list is included in the recording filenames
+        This function compute the  correlation matrix for the cell traces.
 
         Parameters
         ----------
-        strlist : list
-            List of strings to check
-        only_warning: bool, optional
-            If False, throw an exception. Default, True.
+        cellsetname : str
+            cell label to get filename
+        verbose : bool, optional
+            Show additional messages, by default False
 
-         Returns
+        Returns
         -------
-        None
-
+        pd.DataFrame
+            DataFrame with the correlation matrix
         """
-        for x, y in zip(strlist, self.rec_paths):
-            if only_warning and x not in y:
-                print(f"Warning: {x} not in {y}")
-            else:
-                assert x in y, f"Error {x} not in {y}"
+
+        cell_set_files = fh.get_results_filenames(
+            f"{cellsetname}", op=None, single_plane=False
+        )
+        corr_files = fh.get_results_filenames(
+            f"{cellsetname}_corr.csv", op=None, single_plane=False
+        )
+        return compute_traces_corr(cell_set_files, corr_files, verbose=verbose)
 
     def get_results_filenames(
         self,
@@ -554,9 +550,7 @@ class isx_files_handler:
         None
 
         """
-        if op == "DI":
-            paths = self.deinterleave_output_files
-        elif op == "DFF":
+        if op == "DFF":
             paths = self.get_results_filenames("dff", op=None)
         elif op == "PM":
             paths = self.get_results_filenames("maxdff", op=None)
@@ -600,12 +594,7 @@ class isx_files_handler:
         """
 
         if pairlist is None:
-            if op == "MC":
-                translation_files = self.get_results_filenames(
-                    "translations.csv", op=op
-                )
-                crop_rect_files = self.get_results_filenames("crop_rect.csv", op=op)
-            elif op == "DFF":
+            if op == "DFF":
                 input = self.get_filenames(op=None)
                 output = self.get_results_filenames("dff", op=None)
                 amount_of_files = len(input)
@@ -617,20 +606,12 @@ class isx_files_handler:
                 pairlist = zip(input, output)
             if op in ["PP", "TR", "BP", "MC"]:
                 amount_of_files, pairlist = self.get_pair_filenames(op)
-        elif op == "MC":
-            translation_files = [
-                os.path.splitext(p[1])[0] + "-translations.csv" for p in pairlist
-            ]
-            crop_rect_files = [
-                os.path.splitext(p[1])[0] + "-crop_rect.csv" for p in pairlist
-            ]
 
         if overwrite:
             for _, output in pairlist:
                 remove_file_and_json(output)
 
         steps_list = {
-            "DI": "de-interleave",
             "PP": "preprocessing",
             "BP": "spatial_filter",
             "MC": "motion_correct",
@@ -643,53 +624,32 @@ class isx_files_handler:
         assert len(operation) == 1, f"Step {op} not starts with {steps_list}."
         operation = operation[0]
 
-        if op != "DI":
-            parameters = self.default_parameters[operation].copy()
-            for key, value in kws.items():
-                assert key in parameters, f"The parameter: {key} does not exist"
-                parameters[key] = value
+        parameters = self.default_parameters[operation].copy()
+        for key, value in kws.items():
+            assert key in parameters, f"The parameter: {key} does not exist"
+            parameters[key] = value
 
-        # TASK: No need for this step at this point - Can you confirm Fernando?
-        '''
-        if self.processing_steps.index(op) == 0:  # if first step
-            # check if all exist
-            for finput, _ in self.get_pair_filenames(op):
-                assert os.path.exists(
-                    finput
-                ), f"""File f{input} not exist:
-                    Run .de_interleave() to De-interleave multiplane movies"""
-        '''
-
-        if op.startswith("DI"):
-            print("De-interleaving movies, please wait...")
-            self.deinterleave_output_files = de_interleave(
-                focus_files=self.focus_files, efocus=self.efocus
-            )
-        if op.startswith("PP"):
-            print("Preprocessing movies, please wait...")
-            preprocess_step(pairlist, parameters, amount_of_files, verbose)
-        if op.startswith("BP"):
-            print("Applying bandpass filter, please wait...")
-            spatial_filter_step(pairlist, parameters, amount_of_files, verbose)
-        if op.startswith("MC"):
-            print("Applying motion correction, Please wait...")
-            motion_correct_step(
-                translation_files,
-                crop_rect_files,
-                parameters,
-                pairlist,
-                amount_of_files,
-                verbose,
-            )
+        if self.processing_steps.index(op) == 0:  # deinterleave needed
+            pb = progress_bar(len(self.focus_files), "Deinterleaving")
+            for (main_file, planes_fs), focus in zip(
+                self.focus_files.items(), self.efocus
+            ):
+                de_interleave(main_file, planes_fs, focus)
+                pb.update_progress_bar(1)
+        if (
+            op.startswith("PP")
+            or op.startswith("BP")
+            or op.startswith("MC")
+            or op.startswith("DFF")
+            or op.startswith("PM")
+        ):
+            pb = progress_bar(amount_of_files, f_message[operation])
+            for input, output in pairlist:
+                f_register[operation](input, output, parameters, verbose)
+                pb.update_progress_bar(1)
         if op.startswith("TR"):
             print("Trim movies, Please wait...")
             trim_movie(pairlist, parameters, amount_of_files, verbose)
-        if op.startswith("DFF"):
-            print("Normalizing via DF/F0, Please wait...")
-            create_dff(pairlist, parameters, amount_of_files, overwrite, verbose)
-        if op.startswith("PM"):
-            print("Projecting Movie, Please wait...")
-            project_movie(pairlist, parameters, amount_of_files, overwrite, verbose)
         print("done")
 
     @timer
@@ -1285,301 +1245,6 @@ class isx_files_handler:
         print(f"Current total execution time {timedelta(seconds=cls.total_time)}")
 
 
-def de_interleave(focus_files: dict, efocus: list, overwrite: bool = False) -> None:
-    """
-    This function applies the isx.de_interleave function, which de-interleaves multiplane movies
-
-    Parameters
-    ----------
-    overwrite : Bool, optional
-        If True the function erases the previous information; otherwise, it appends to a list.
-        By default "False"
-
-    Returns
-    -------
-    None
-
-    """
-
-    # Initialize progress bar
-    pb = progress_bar(len(focus_files), "Deinterleaving")
-    output_files = []
-    for (main_file, planes_fs), focus in zip(focus_files.items(), efocus):
-        if len(focus) > 1:  # has multiplane
-            existing_files = []
-            for sp_file in planes_fs:
-                dirname = os.path.dirname(sp_file)
-                json_file = os.path.splitext(sp_file)[0] + ".json"
-                if os.path.exists(sp_file):
-                    if overwrite:
-                        os.remove(sp_file)
-                        if os.path.exists(json_file):
-                            os.remove(json_file)
-                    else:
-                        if same_json_or_remove(
-                            parameters={
-                                "input_movie_files": main_file,
-                                "output_movie_files": [
-                                    os.path.basename(p) for p in planes_fs
-                                ],
-                                "in_efocus_values": focus,
-                            },
-                            input_files_keys=["input_movie_files"],
-                            output=sp_file,
-                            verbose=False,
-                        ):
-                            existing_files.append(sp_file)
-            if len(existing_files) != len(planes_fs):  # has files to run
-                for f in existing_files:  # remove existing planes
-                    os.remove(f)
-                try:
-                    isx.de_interleave(main_file, planes_fs, focus)
-
-                except Exception as err:
-                    print("Reading: ", main_file)
-                    print("Writting: ", planes_fs)
-                    raise err
-
-            data = {
-                "input_movie_files": main_file,
-                "output_movie_files": [os.path.basename(p) for p in planes_fs],
-                "in_efocus_values": focus,
-            }
-            # for sp_file in planes_fs:
-            write_log_file(
-                params=data,
-                dir_name=dirname,
-                input_files_keys=["input_movie_files"],
-                output_file_key="output_movie_files",
-            )
-
-        # Update progress bar
-        pb.update_progress_bar(1)
-        output_files += planes_fs
-    # Need to return value to save the output values back in the object for the remove function later used.
-    return output_files
-
-
-def preprocess_step(
-    pairlist: Tuple[list, list],
-    parameters: dict,
-    amount_of_files: int,
-    verbose: bool = False,
-) -> None:
-    """
-    After performing checks, use the isx.preprocess function, which preprocesses movies,
-    optionally applying spatial and temporal downsampling and cropping.
-
-    Parameters
-    ----------
-    pairlist : Tuple[list, list]
-        Tuple containing lists of input and output paths
-    parameters : dict
-       Parameter dictionary of executed functions.
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    """
-
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Preprocessing")
-
-    for input, output in pairlist:
-        if isinstance(parameters["spatial_downsample_factor"], str):
-            res_idx, value = parameters["spatial_downsample_factor"].split("_")
-            if res_idx == "maxHeight":
-                idx_resolution = 0
-            elif res_idx == "maxWidth":
-                idx_resolution = 1
-            else:
-                assert False, "error in sp_downsampling parameter value"
-            movie = isx.Movie.read(input)
-            resolution = movie.spacing.num_pixels
-            del movie
-            parameters["spatial_downsample_factor"] = np.ceil(
-                resolution[idx_resolution] / float(value)
-            )
-
-        parameters.update(
-            {
-                "input_movie_files": os.path.basename(input),
-                "output_movie_files": os.path.basename(output),
-            }
-        )
-
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-
-        isx.preprocess(
-            **parameters_for_isx(
-                parameters,
-                ["comments", "fix_frames_th_std"],
-                {"input_movie_files": input, "output_movie_files": output},
-            )
-        )
-
-        nfixed = fix_frames(output, std_th=parameters["fix_frames_th_std"], report=True)
-
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "preprocess"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        if verbose:
-            print("{} preprocessing completed. {} frames fixed.".format(output, nfixed))
-        # Update progress bar
-        pb.update_progress_bar(1)
-
-
-def spatial_filter_step(
-    pairlist: Tuple[list, list],
-    parameters: dict,
-    amount_of_files: int,
-    verbose: bool = False,
-) -> None:
-    """
-    After performing checks, use the isx.spatial_filter function, which
-    apply spatial bandpass filtering to each frame of one or more movies
-
-    Parameters
-    ----------
-    pairlist: Tuple[list, list]
-        Tuple containing lists of input and output paths
-    parameters : dict
-        Parameter dictionary of executed functions.
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    """
-
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Applying Bandpass Filter to")
-
-    for input, output in pairlist:
-        parameters.update(
-            {
-                "input_movie_files": os.path.basename(input),
-                "output_movie_files": os.path.basename(output),
-            }
-        )
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-        isx.spatial_filter(
-            **parameters_for_isx(
-                parameters,
-                ["comments"],
-                {"input_movie_files": input, "output_movie_files": output},
-            )
-        )
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "spatial_filter"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        if verbose:
-            print("{} bandpass filtering completed".format(output))
-        # Update progress bar
-        pb.update_progress_bar(1)
-
-
-def motion_correct_step(
-    translation_files: list,
-    crop_rect_files: list,
-    parameters: dict,
-    pairlist: Tuple[list, list],
-    amount_of_files: int,
-    verbose=False,
-) -> None:
-    """
-    After checks, use the isx.motion_correct function, which motion correct movies to a reference frame.
-
-    Parameters
-    ----------
-    translation_files : list
-        A list of file names to write the X and Y translations to. Must be either None,
-        in which case no files are written, or a list of valid file names equal in
-        length to the number of input and output file names
-    crop_rect_files : list
-        The path to a file that will contain the crop rectangle applied to the input
-        movies to generate the output movies
-    parameters : dict
-       Parameter list of executed functions.
-    pairlist : Tuple[list, list]
-        Tuple containing lists of input and output paths
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    """
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Applying Motion Correction to")
-    for i, (input, output) in enumerate(pairlist):
-        new_data = {
-            "input_movie_files": os.path.basename(input),
-            "output_movie_files": os.path.basename(output),
-            "output_translation_files": os.path.basename(translation_files[i]),
-            "output_crop_rect_file": os.path.basename(crop_rect_files[i]),
-        }
-        parameters.update(new_data)
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-        isx.motion_correct(
-            **parameters_for_isx(
-                parameters,
-                ["comments"],
-                {
-                    "input_movie_files": input,
-                    "output_movie_files": output,
-                    "output_translation_files": translation_files[i],
-                    "output_crop_rect_file": crop_rect_files[i],
-                },
-            )
-        )
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "motion_correct"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        if verbose:
-            print("{} motion correction completed".format(output))
-        # Update progress bar
-        pb.update_progress_bar(1)
-
-
 def trim_movie(
     pairlist: Tuple[list, list],
     user_parameters: dict,
@@ -1650,264 +1315,3 @@ def trim_movie(
         )
         # Update progress bar
         pb.update_progress_bar(1)
-
-
-def create_dff(
-    pairlist: Tuple[list, list],
-    parameters: dict,
-    amount_of_files: int,
-    overwrite=False,
-    verbose=False,
-    **kws,
-) -> None:
-    """
-    This function applies isx.dff function, to compute the DF/F movies.
-
-    Parameters
-    ----------
-    overwrite : bool, optional
-        Remove results and recompute them, by default False
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    """
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Normalizing via DF/F0")
-
-    if overwrite:
-        for input, output in zip(pairlist):
-            remove_file_and_json(output)
-
-    for key, value in kws.items():
-        assert key in parameters, f"The parameter: {key} does not exist"
-        parameters[key] = value
-
-    for input, output in pairlist:
-        new_data = {
-            "input_movie_files": os.path.basename(input),
-            "output_movie_files": os.path.basename(output),
-        }
-        parameters.update(new_data)
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-        isx.dff(
-            **parameters_for_isx(
-                parameters,
-                ["comments"],
-                {"input_movie_files": input, "output_movie_files": output},
-            )
-        )
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "dff"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_movie_files",
-        )
-        # Update progress bar
-        pb.update_progress_bar(1)
-    print("done")
-
-
-def project_movie(
-    pairlist: Tuple[list, list],
-    parameters: dict,
-    amount_of_files: int,
-    overwrite=False,
-    verbose=False,
-    **kws,
-) -> None:
-    """
-    This function applies isx.project_movie to project movies to a single statistic image.
-
-    Parameters
-    ----------
-    input_name : str, optional
-        Input file path, by default "dff"
-    output_name : str, optional
-        Output file path, by default "maxdff"
-    operation : str, optional
-        Preprocessing operation to check, by default None.
-    overwrite : bool, optional
-        Remove results and recompute them, by default False
-    verbose : bool, optional
-        Show additional messages, by default False
-
-    Returns
-    -------
-    None
-
-    """
-
-    # Initialize progress bar
-    pb = progress_bar(amount_of_files, "Projecting Movies")
-
-    if overwrite:
-        # get rid of self
-        for input, output in pairlist:
-            remove_file_and_json(output)
-
-    for key, value in kws.items():
-        assert key in parameters, f"The parameter: {key} does not exist"
-        parameters[key] = value
-
-    # turn into pairlist
-    for input, output in pairlist:
-        new_data = {
-            "input_movie_files": os.path.basename(input),
-            "output_image_file": os.path.basename(output),
-        }
-        parameters.update(new_data)
-        if same_json_or_remove(
-            parameters,
-            input_files_keys=["input_movie_files"],
-            output=output,
-            verbose=verbose,
-        ):
-            continue
-        isx.project_movie(
-            **parameters_for_isx(
-                parameters,
-                ["comments"],
-                {"input_movie_files": input, "output_image_file": output},
-            )
-        )
-        write_log_file(
-            parameters,
-            os.path.dirname(output),
-            {"function": "project_movie"},
-            input_files_keys=["input_movie_files"],
-            output_file_key="output_image_file",
-        )
-        # Update progress bar
-        pb.update_progress_bar(1)
-    print("done")
-
-
-def get_segment_from_movie(
-    inputfile: str,
-    outputfile: str,
-    borders: Iterable,
-    keep_start_time=False,
-    unit="minutes",
-) -> None:
-    """
-    This function gets a segment of a video to create a new one. It's just an easy
-    handler of isx.trim_movie, which trim frames from a movie to produce a new movie
-
-    Parameters
-    ----------
-    inputfile : str
-        Path of the input file to use
-    outputfile : str
-        Path of the output file to use
-    borders: iterable
-        With two elements of the borders in minutes (integers) of the segment.
-        Where negative times means from the end like indexing in a numpy array
-    keep_start_time : Bool, optional
-        If true, keep the start time of the movie, even if some of its
-        initial frames are to be trimmed, by default False
-    unit: str
-        the unit to use for border. It could be 'minutes' or 'seconds'
-
-    Returns
-    -------
-    None
-
-    Examples
-    --------
-    >>> get_segment_from_movie('inputfile.isxd','output.isxd',[-30, -1]) #last 30 minutes
-
-    """
-    assert len(borders) == 2, "borders must have two elements"
-    assert unit in ["minutes", "seconds"], """unit could be 'minutes' or 'seconds'. """
-    movie = isx.Movie.read(inputfile)
-    if unit == "minutes":
-        unitfactor = 60
-    else:
-        unitfactor = 1
-    numframes = movie.timing.num_samples
-    duration_unit = numframes * movie.timing.period.to_msecs() / (1000 * unitfactor)
-    timing_period_unit = movie.timing.period.to_msecs() / (1000 * unitfactor)
-
-    movie.flush()
-
-    crop_segments = []
-    assert (
-        borders[0] > -duration_unit
-    ), "asking for a segment from {minutes[0]} {unit} before the end, but video is {duration_minutes} {unit} long."
-    assert (
-        borders[0] < duration_unit
-    ), "asking for a segment from {minutes[0]} {unit}, but video is {duration_minutes} {unit} long."
-    assert (
-        borders[1] < duration_unit
-    ), "asking for a segment up to {minutes[0]} {unit}, but video is {duration_minutes} {unit} long."
-    assert (
-        borders[1] > -duration_unit
-    ), "asking for a segment up to {minutes[0]} {unit} before the end, but video is {duration_minutes} {unit} long."
-
-    # remove fist frames:
-    # don't cut if the segments are from the beggining or just exactlt
-    # duration before the end
-    if borders[0] != 0 and borders[0] != -duration_unit:
-        if borders[0] < 0:
-            end1 = (duration_unit + borders[0]) / timing_period_unit - 1
-        else:
-            end1 = borders[0] / timing_period_unit - 1
-        crop_segments.append([0, int(end1)])
-    # remove last frames:
-    # and don't cut if the segments are up to the end or just the exact
-    # duration
-    if borders[1] != -1 and borders[1] != duration_unit:
-        if borders[1] < 0:
-            start1 = (duration_unit + borders[1] + 1) / timing_period_unit + 1
-        else:
-            start1 = borders[1] / timing_period_unit + 1
-        crop_segments.append([int(start1), numframes])
-
-    if os.path.exists(outputfile):
-        os.remove(outputfile)
-    if len(crop_segments) == 0:
-        print("no trim need it")
-        return
-    isx.trim_movie(
-        input_movie_file=inputfile,
-        output_movie_file=outputfile,
-        crop_segments=np.array(crop_segments),
-        keep_start_time=keep_start_time,
-    )
-
-
-def get_efocus(gpio_file: str) -> list:
-    """
-    Read the gpio set from a file and get the data associated.
-
-    Parameters
-    ----------
-    gpio_file : str
-        path
-
-    Returns
-    -------
-    list
-        list with the video_efocus
-
-    """
-    gpio_set = isx.GpioSet.read(gpio_file)
-    efocus_values = gpio_set.get_channel_data(gpio_set.channel_dict["e-focus"])[1]
-    efocus_values, efocus_counts = np.unique(efocus_values, return_counts=True)
-    min_frames_per_efocus = 100
-    video_efocus = efocus_values[efocus_counts >= min_frames_per_efocus]
-    assert (
-        video_efocus.shape[0] < 4
-    ), f"{gpio_file}: Too many efocus detected, early frames issue."
-    return [int(v) for v in video_efocus]
